Server: -
Client:
fixed Movement bugs
Changed the Castle to a Spirte

+++ Code/LordOfWar/core/src/main/java/code/lordofwar/backend/PathCell.java
@@ -4,11 +4,11 @@ import com.badlogic.gdx.math.Vector2;
 
 public class PathCell {
     public Vector2 coords;
-    public int[] distances;
+    public double[] distances;
     //use this value to get most efficient path after arriving at the goal
     public PathCell parent;
 
-    public PathCell(Vector2 coords, int[] distances, PathCell parent) {
+    public PathCell(Vector2 coords, double[] distances, PathCell parent) {
         this.coords = coords;
         this.distances = distances;
         this.parent = parent;
